+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- AccountプロダクトタイプをPaymentCardに統合
+
 ### Deprecated
 
 ### Removed